move etomica.graph test classes to separate source set
