+++ etomica-graph/src/test/java/etomica/graph/ConnectivityTest.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package etomica.graph.test;
+package etomica.graph;
 
 import static etomica.graph.model.Metadata.COLOR_CODE_0;
 import static etomica.graph.model.Metadata.COLOR_CODE_1;
@@ -11,14 +11,8 @@ import java.util.HashMap;
 import java.util.Map;
 
 
-import etomica.graph.iterators.DefaultIterator;
 import etomica.graph.iterators.PartitionedIterator;
-import etomica.graph.iterators.filters.IsomorphismFilter;
-import etomica.graph.iterators.filters.PropertyFilter;
 import etomica.graph.model.GraphIterator;
-import etomica.graph.property.HasNoRootEdge;
-import etomica.graph.property.IsBiconnected;
-import etomica.graph.property.IsConnected;
 
 public class ConnectivityTest extends GraphIteratorTest {
