+++ etomica-graph/src/test/java/etomica/graph/DeleteTest.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package etomica.graph.test;
+package etomica.graph;
 
 import static etomica.graph.model.Metadata.COLOR_CODE_0;
 import static etomica.graph.model.Metadata.COLOR_CODE_1;
@@ -11,12 +11,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
-import etomica.graph.iterators.DefaultIterator;
-import etomica.graph.iterators.IsomorphismPrefilteredPartitionedIterator;
 import etomica.graph.iterators.IteratorToSet;
-import etomica.graph.iterators.PartitionedIterator;
-import etomica.graph.iterators.StoredIterator;
-import etomica.graph.iterators.filters.IsomorphismFilter;
 import etomica.graph.model.Graph;
 import etomica.graph.model.GraphIterator;
 import etomica.graph.operations.Binary;
