+++ etomica-graph/src/test/java/etomica/graph/EdgeCountFilterTest.java
@@ -2,13 +2,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package etomica.graph.test;
+package etomica.graph;
 
 
-import etomica.graph.iterators.FixedEdgeCountIterator;
+import etomica.graph.iterators.DefaultIterator;
+import etomica.graph.iterators.filters.PropertyFilter;
 import etomica.graph.model.GraphIterator;
+import etomica.graph.property.HasEdgeCount;
 
-public class FixedEdgeCountIteratorTest extends GraphIteratorTest {
+public class EdgeCountFilterTest extends GraphIteratorTest {
 
   protected void testNaive(byte nodeCount, GraphIterator iterator) {
 
@@ -23,10 +25,10 @@ public class FixedEdgeCountIteratorTest extends GraphIteratorTest {
     checkAssertion = true;
   }
 
-  public void testEdgeCountIterator() {
+  public void testEdgeFilteredGraphs() {
 
     reset();
-    // nodes = 7: total of 2097152 graphs after 7 secs (0 min)
+    // nodes = 7: total of 2097152 graphs after 56 secs (0 min)
     byte nodes = 7;
     checkAssertion = false;
     int count = 0;
@@ -34,7 +36,7 @@ public class FixedEdgeCountIteratorTest extends GraphIteratorTest {
     for (byte edges = 0; edges <= nodes * (nodes - 1) / 2; edges++) {
       System.out.println("edges....: " + edges);
       System.out.println("partial..: " + count);
-      testNaive(nodes, new FixedEdgeCountIterator(nodes, edges));
+      testNaive(nodes, new PropertyFilter(new DefaultIterator(nodes), new HasEdgeCount(edges)));
       count += enumerated;
     }
     long elapsed = (System.nanoTime() - time) / 1000000000;
