+++ etomica-graph/src/test/java/etomica/graph/FixedEdgeCountIteratorTest.java
@@ -2,15 +2,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package etomica.graph.test;
+package etomica.graph;
 
 
-import etomica.graph.iterators.DefaultIterator;
-import etomica.graph.iterators.filters.PropertyFilter;
+import etomica.graph.iterators.FixedEdgeCountIterator;
 import etomica.graph.model.GraphIterator;
-import etomica.graph.property.HasEdgeCount;
 
-public class EdgeCountFilterTest extends GraphIteratorTest {
+public class FixedEdgeCountIteratorTest extends GraphIteratorTest {
 
   protected void testNaive(byte nodeCount, GraphIterator iterator) {
 
@@ -25,10 +23,10 @@ public class EdgeCountFilterTest extends GraphIteratorTest {
     checkAssertion = true;
   }
 
-  public void testEdgeFilteredGraphs() {
+  public void testEdgeCountIterator() {
 
     reset();
-    // nodes = 7: total of 2097152 graphs after 56 secs (0 min)
+    // nodes = 7: total of 2097152 graphs after 7 secs (0 min)
     byte nodes = 7;
     checkAssertion = false;
     int count = 0;
@@ -36,7 +34,7 @@ public class EdgeCountFilterTest extends GraphIteratorTest {
     for (byte edges = 0; edges <= nodes * (nodes - 1) / 2; edges++) {
       System.out.println("edges....: " + edges);
       System.out.println("partial..: " + count);
-      testNaive(nodes, new PropertyFilter(new DefaultIterator(nodes), new HasEdgeCount(edges)));
+      testNaive(nodes, new FixedEdgeCountIterator(nodes, edges));
       count += enumerated;
     }
     long elapsed = (System.nanoTime() - time) / 1000000000;
