+++ etomica-graph/src/test/java/etomica/graph/GraphIteratorTest.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package etomica.graph.test;
+package etomica.graph;
 
 import java.util.Iterator;
